Refactor e2e pod management utilities
Moves pod creation logic from a global utility to a method on the `Client` struct, enabling address initialization during creation.

Introduces a dedicated `Pod` struct to encapsulate WireGuard key management, network addresses, and CNI configuration generation logic. This centralizes pod-specific state and behavior.

Adds helper methods like `ClientPeer` and `ServerPeer` on the `Pod` struct to simplify the creation of CNI peer configurations within tests.

Streamlines the `CniConfig` method on `Pod` by leveraging its internal address, removing the need to pass it explicitly in test calls.

Simplifies the e2e Kubernetes test setup by providing a more object-oriented and encapsulated approach to managing test pods.
